requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alluka==0.4.0
attrs==24.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
colorlog==6.9.0
frozenlist==1.5.0
hikari==2.1.0
hikari-tanjun==2.17.7
idna==3.10
multidict==6.1.0
propcache==0.2.1
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.8.0
typing_extensions==4.12.2
urllib3==2.3.0
wheel==0.45.1
yarl==1.18.3
